+++ prototype_kl/parser/config.go
@@ -112,6 +112,7 @@ var SupportedFunctions = map[string]int{
 	"hypot":     2,  // hypot(x, y)
 	"root":      2,
 	"sum":       2,
+	"fix":       2,
 	"prod":      2,
 	"min":       -1, // variadic
 	"max":       -1, // variadic